+++ lib/anycable/middleware.rb
@@ -3,7 +3,7 @@
 module AnyCable
   # Middleware is an analague of Rack middlewares but for AnyCable RPC calls
   class Middleware
-    def call(_method_name, _request)
+    def call(_method_name, _request, _meta)
       raise NotImplementedError
     end
   end