+++ spec/integrations/grpc/connection_spec.rb
@@ -32,7 +32,7 @@ describe "client connection" do
     allow(AnyCable.rpc_handler).to receive(:handle).and_call_original
 
     expect(subject).to be_success
-    expect(AnyCable.rpc_handler).to have_received(:handle).with(:connect, request)
+    expect(AnyCable.rpc_handler).to have_received(:handle).with(:connect, request, an_instance_of(Hash))
   end
 
   context "when exception" do