+++ spec/integrations/grpc/disconnection_spec.rb
@@ -33,7 +33,7 @@ describe "disconnection" do
     allow(AnyCable.rpc_handler).to receive(:handle).and_call_original
 
     expect(subject).to be_success
-    expect(AnyCable.rpc_handler).to have_received(:handle).with(:disconnect, request)
+    expect(AnyCable.rpc_handler).to have_received(:handle).with(:disconnect, request, an_instance_of(Hash))
   end
 
   context "when exception" do