+++ spec/integrations/grpc/perform_spec.rb
@@ -22,7 +22,7 @@ describe "client messages" do
     allow(AnyCable.rpc_handler).to receive(:handle).and_call_original
 
     expect(subject).to be_success
-    expect(AnyCable.rpc_handler).to have_received(:handle).with(:command, request)
+    expect(AnyCable.rpc_handler).to have_received(:handle).with(:command, request, an_instance_of(Hash))
   end
 
   context "when exception" do